Allow resume training from checkpoint
In the previous version of the library training from a checkpoint was
not possible, because always the latest version in the experiment
directory was fetched for resume training. This was adapted such that to
resume a specific checkpoint should be added. The difference between
train from checkpoint and resume from checkpoint is then whether the
optimizer values are kept.