Fix vocab loading in train_model
In the previous version of the library the vocabulary was always built
from the training set, resulting in a potential mismatch with the
vocabulary of a trained model. Fix this by setting the scr and tgt field
vocabularies to the vocabularies of the loaded checkpoint.

TODO: A test should be written to cover this.