Include post and pre decoder rnn attention
Change attention mechanism such that both applying it before and
applying it after the decoder rnn is available. This slightly changes
the api of the train_model.py file: instead of using --attention as a
mere flag, it now requires an argument that specifies if the attention
should be used pre- or post decoder.