Remove duplicate functionality
Methods to compute the loss of a batch were present in both the
evaluator and the trainer, requiring changes at two different timesteps
if a change in the loss computation is made.
Now, to compute the loss in the trainer, the `compute_batch_loss`
function in the evaluator is called, hereby getting rid of duplicate
lines of code.