Add new meric to evaluate last step of computation
For models that ponder before then presenting the final output, it makes
sense to have a metric that computes the accuracy only on this final
output. The new metric `FinalTargetAccuracy` computes this accuracy, by
taking into account only the last value in the sequence (i.e. the one
before <eos>, or before the first padding in case <eos> is not included
in the target).