Correct loop in FinalTargetAccuracy computation
The loop in the FinalTargetAccuracy did not loop up until the last
element of the sequence. Because before the last element of the sequence
was always <eos> this was not a problem, but without <eos> the actual
target was often never reached, resulting in a FinalTargetAccuracy of 0.
The loop is now fixed.