Addition to e4a8a17
In batches we should also 'remove' the last attention index for shorter sequences by setting them to -1
This makes sure that we won't get NaN issues in the hard attention method.
What otherwise would happen is that we would attend to <pad> inputs. This would be set to -inf by the mask, resulting in a softmax over only -inf's